Release version 2.3.33
Source: b8f133b99f4d11772c510d80b8617ecd97c9f524

Reviewed-on: https://gerrit.bouffalolab.com/c/release/bouffalo_sdk/+/10817
Tested-by: zrrong <zrrong@bouffalolab.com>

+++ components/fota/https/https_fota.c
@@ -74,7 +74,7 @@ static int https_fota_cancel_requested(void *user_data)
     return (fota != NULL && fota->abort_requested) ? 1 : 0;
 }
 
-static int __attribute__((unused)) payload_cb(int sock, struct http_request *req, void *user_data)
+static int __attribute__((unused)) payload_cb(intptr_t sock, struct http_request *req, void *user_data)
 {
     const char *content[] = {
         "foobar",
